Remove Errors
This includes linting and unused imports
